requirements.txt
# Plant3D Python 패키지 의존성
# 경로: E:\github\plant3D\requirements.txt

# Streamlit 웹앱
streamlit==1.28.0
plotly==5.17.0

# PDF 처리
PyPDF2==3.0.1
tabula-py==2.8.2
pdfplumber==0.10.3
pillow==10.1.0

# 데이터 처리
pandas==2.1.4
numpy==1.26.2
pyyaml==6.0.1
jsonschema==4.20.0

# 웹 서버
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
gunicorn==21.2.0

# 프로세스 관리
psutil==5.9.6
subprocess32==3.5.4

# 3D 파일 처리
trimesh==4.0.8
numpy-stl==3.0.1

# 로깅 및 모니터링
loguru==0.7.2
tqdm==4.66.1
colorama==0.4.6

# 개발 도구
pytest==7.4.3
black==23.12.1
flake8==7.0.0

# API 통신
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# 파일 시스템 감시
watchdog==3.0.0

# 데이터 검증
marshmallow==3.20.1
python-dotenv==1.0.0